Initial commit.
This is an extraction of some of our natural
language processing tools code.

At this point the primary function is to be an
interface in Ruby to the Stanford CoreNLP suite
of tools. The interface takes the XML output files
from the parser, and builds Ruby objects from them,
for use within a Ruby application.

The gem also provides a rake task to download
the Stanford CoreNLP dependencies files, so
developers and our production server all use the
same version, and we can keep the files out of
version control.